feat: add signup controller
